pyproject.toml
[tool.poetry]
name = "hmw5"
version = "0.1.0"
description = ""
authors = ["Maria Scofa <mariascofa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.1.4"
flake8 = "3.8.4"
Faker = "5.3.0"
gender-guesser = "0.4.0"
django-crispy-forms = "1.10.0"
celery = "5.0.5"
SQLAlchemy = "1.3.22"
django-extensions = "3.1.0"
django-filter = "2.4.0"
requests = "2.25.1"
ephem = "3.7.7"
django-redis = "4.12.1"
gunicorn = "20.0.4"
Pillow = "8.1.2"

[tool.poetry.dev-dependencies]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
